requirements.txt
flask
flask_mail
flask_sqlalchemy
flask_login
Faker
dotenv
Pillow
psycopg2-binary
requests
